requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
chromadb==0.5.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pandas==2.2.2
openai==1.45.0
python-dotenv==1.0.1
google-auth==2.34.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
PyMuPDF==1.24.9
google-auth-oauthlib